Rename src/ to lib/
